refactor: split asset proxy pwa modules
